% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_summary.R
\name{class_summary}
\alias{class_summary}
\title{class_summary}
\usage{
class_summary(data)
}
\arguments{
\item{data}{Requires a  data frame generated by score_genes; class - data frame}
}
\value{
multiple plots; class - plots
}
\description{
class_summary function provides information on main protein classes. Barplot also helps to visualise  the class distribution. Function depends on  the data frame generated by score_genes function,
}
\examples{
\dontrun{
path_to_test_data<- system.file("extdata", "test_data.tabular", package="OmicInt")
# basic usage of class_summary
df<-utils::read.table(path_to_test_data)
class_summary(df)}
}
