% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{density_plot}
\usage{
density_plot(data)
}
\arguments{
\item{data}{Requires a  data frame generated by score_genes; class - data frame}
}
\value{
multiple plots; class - plots
}
\description{
Function plots a density plot for gene expression data prepared by the score_genes function. The plots can be used for a quick assessment of the overall gene expression distribution.
}
\examples{
\dontrun{
path_to_test_data<- system.file("extdata", "test_data.tabular", package="OmicInt")
# basic usage of density_plot
df<-utils::read.table(path_to_test_data)
density_plot(df)}
}
