% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_distribution.R
\name{feature_distribution}
\alias{feature_distribution}
\title{feature_distribution}
\usage{
feature_distribution(data)
}
\arguments{
\item{data}{Requires a  data frame generated by score_genes function; class - data frame}
}
\value{
multiple summary plots; class - plots
}
\description{
Function collects data from STRING database, scales, as well as prepares additional score integration to visualise data feature distribution.
}
\examples{
 \dontrun{
path_to_test_data<- system.file("extdata", "test_data.tabular", package="OmicInt")
# basic usage of feature_distribution
df<-utils::read.table(path_to_test_data)
feature_distribution(df)}
}
