% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK_genes.R
\name{HK_genes}
\alias{HK_genes}
\title{HK_genes}
\usage{
HK_genes(data, meta)
}
\arguments{
\item{data}{Requires a  data frame of normalised scores in CSV format}

\item{meta}{Requires a  data frame of metadata  in CSV format}
}
\value{
multiple plots
}
\description{
HK_genes function provides a way to visualise how housekeeping genes changed throughout the conditions under the investigation. Depending on the number of conditions separate plots will be generated.
}
\examples{
path_to_test_data<- system.file("extdata", "normalised_counts.csv", package="OmicInt")
path_to_meta_data<- system.file("extdata", "metadata.csv", package="OmicInt")
# basic usage of HK_genes
HK_genes(path_to_test_data,path_to_meta_data)
}
