% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactor_map.R
\name{interactor_map}
\alias{interactor_map}
\title{interactor_map}
\usage{
interactor_map(data)
}
\arguments{
\item{data}{requires a data frame containing gene names as row names and a column with LFC values; class - data frame}
}
\value{
interaction map/plot; class - plot
}
\description{
interactor_map uses information mined from STRING database to map  experimental, predicted,  or referenced interactions  to see if there are  any interactors in the set of significantly changed genes and how they are linked. The function requires a data frame prepared by score_genes. The output is a plot depicting interaction map.
}
\examples{
 \dontrun{
path_to_test_data<- system.file("extdata", "test_data.tabular", package="OmicInt")
# basic usage of interactor_map
df<-utils::read.table(path_to_test_data)
interactor_map(df)
}
}
