% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_summary.R
\name{location_summary}
\alias{location_summary}
\title{location_summary}
\usage{
location_summary(data)
}
\arguments{
\item{data}{Requires a  data frame generated by score_genes; class - data frame}
}
\value{
barplot; class - plot
}
\description{
location_summary  function provides information on main cellular locations. Barplot also helps to visualise  the location data distribution.
}
\examples{
\dontrun{
path_to_test_data<- system.file("extdata", "test_data.tabular", package="OmicInt")
# basic usage of location_summary
df<-utils::read.table(path_to_test_data)
location_summary(df)}
}
