% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_search.R
\name{pattern_search}
\alias{pattern_search}
\title{pattern_search}
\usage{
pattern_search(data, meta, Condition = "Condition_1")
}
\arguments{
\item{data}{Requires a  data frame of normalised scores in CSV format}

\item{meta}{Requires a  data frame of metadata  in CSV format}

\item{Condition}{Requires a condition to select if multiple, default "Condition_1"}
}
\value{
pattern list
}
\description{
pattern_search function searches for gene patterns that were upregulated or downregulated throughout the conditions when comparing to the geometric mean across all conditions. The geometric mean serves as a base value to compare across multiple conditions if more complex patterns exist and also allows for a universal baseline.
}
\examples{
path_to_test_data<- system.file("extdata", "normalised_counts.csv", package="OmicInt")
path_to_meta_data<- system.file("extdata", "metadata.csv", package="OmicInt")
# basic usage of pattern_search
pattern_search(path_to_test_data,path_to_meta_data)
}
