% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_map.R
\name{class_map}
\alias{class_map}
\title{class_map}
\usage{
class_map(data, num = 0)
}
\arguments{
\item{data}{Requires a  data frame generated by score_genes; class - data frame}

\item{num}{a number for genes to cluster from top upregulated and downregulated genes, if not selected all genes will be used; default 0, i.e. do not select a smaller set; class - integer}
}
\value{
dendogram; class - plot
}
\description{
Function provides visualisation of how the highest and lowest LFCscore  genes cluster based on protein class data. Function requires a data frame generated by the score_genes function.
}
\examples{
\dontrun{
path_to_test_data<- system.file("extdata", "test_data.tabular", package="OmicInt")
# basic usage of class_map
df<-utils::read.table(path_to_test_data)
class_map(df)}
}
