% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HK_genes.R
\name{HK_genes}
\alias{HK_genes}
\title{HK_genes}
\usage{
HK_genes(data, meta)
}
\arguments{
\item{data}{Requires a path variable to a data file of normalised scores in CSV format (comma separated); class - string}

\item{meta}{Requires a path variable to a data file of metadata  in CSV format (comma separated); class - string}
}
\value{
multiple plots; class - plots
}
\description{
HK_genes function provides a way to visualise how housekeeping genes changed throughout the conditions under the investigation. Depending on the number of conditions separate plots will be generated. Function requires a path variable to a normalised count data file.
}
\examples{
\dontrun{
path_to_test_data<- system.file("extdata", "normalised_counts.csv", package="OmicInt")
path_to_meta_data<- system.file("extdata", "meta_data.csv", package="OmicInt")
# basic usage of HK_genes
HK_genes(path_to_test_data,path_to_meta_data)}
}
