% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_plots.R
\name{pattern_plots}
\alias{pattern_plots}
\title{pattern_plots}
\usage{
pattern_plots(data, meta, low = NA, high = NA)
}
\arguments{
\item{data}{Requires a  data frame of normalised scores subsetted from pattern_search function}

\item{meta}{Requires a  data frame of metadata  in CSV format}

\item{low}{the lowest value for the expression value}

\item{high}{the highest value for the expression value}
}
\value{
multiple plots
}
\description{
pattern_plots function uses a subsetted pattern data from the function pattern_search. The function plots distribution plots as well as a selected set of genes and how they changed patterns.
}
\examples{
path_to_test_data<- system.file("extdata", "normalised_counts.csv", package="OmicInt")
path_to_meta_data<- system.file("extdata", "metadata.csv", package="OmicInt")
# basic usage of pattern_search
pattern_search(path_to_test_data,path_to_meta_data)
}
