% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3D_distribution.R
\name{plot_3D_distribution}
\alias{plot_3D_distribution}
\title{plot_3D_distribution}
\usage{
plot_3D_distribution(data, type = "association")
}
\arguments{
\item{data}{a data frame containing processed expression file from score_genes with LFCscore; class - data frame}

\item{type}{default value is "association", the user can select how to color data points depending on association or specificity score (e.g., selecting "specificity"); class - string}
}
\value{
function returns an interactive plot; class - plot
}
\description{
Function allows to explore 3D distribution between the number of interactors, LFCscore and p.adj values. Function takes a data frame provided by score_genes function.
}
\examples{
\dontrun{
path_to_test_data<- system.file("extdata", "test_data.tabular", package="OmicInt")
# basic usage of plot_3D_distribution
df<-utils::read.table(path_to_test_data)
plot_3D_distribution(df)}
}
